refactor: rename program_json.go to encoding.go and implement JSON marshaling/unmarshaling for Program struct
