+++ pkg/bytecode/encoding.go
@@ -24,7 +24,7 @@ type (
 		Registers  int            `json:"registers"`
 		Bytecode   []Instruction  `json:"bytecode"`
 		Constants  []constantJSON `json:"constants,omitempty"`
-		CatchTable []Catch        `json:"catch_table,omitempty"`
+		CatchTable []Catch        `json:"catchTable,omitempty"`
 		Params     []string       `json:"params,omitempty"`
 		Metadata   Metadata       `json:"metadata"`
 	}
@@ -34,67 +34,6 @@ type (
 		Value stdjson.RawMessage `json:"value,omitempty"`
 	}
 )
-
-func (p *Program) MarshalJSON() ([]byte, error) {
-	if p == nil {
-		return []byte("null"), nil
-	}
-
-	constants := make([]constantJSON, len(p.Constants))
-
-	for i, value := range p.Constants {
-		encoded, err := encodeConstant(value)
-
-		if err != nil {
-			return nil, fmt.Errorf("bytecode.Program: encode constant %d: %w", i, err)
-		}
-
-		constants[i] = encoded
-	}
-
-	payload := programJSON{
-		Source:     p.Source,
-		Registers:  p.Registers,
-		Bytecode:   p.Bytecode,
-		Constants:  constants,
-		CatchTable: p.CatchTable,
-		Params:     p.Params,
-		Metadata:   p.Metadata,
-	}
-
-	return json.Marshal(payload)
-}
-
-func (p *Program) UnmarshalJSON(data []byte) error {
-	if p == nil {
-		return fmt.Errorf("bytecode.Program: UnmarshalJSON on nil pointer")
-	}
-
-	var decoded programJSON
-	if err := json.Unmarshal(data, &decoded); err != nil {
-		return err
-	}
-
-	constants := make([]runtime.Value, len(decoded.Constants))
-	for i, value := range decoded.Constants {
-		decodedValue, err := decodeConstant(value)
-		if err != nil {
-			return fmt.Errorf("bytecode.Program: decode constant %d: %w", i, err)
-		}
-
-		constants[i] = decodedValue
-	}
-
-	p.Source = decoded.Source
-	p.Registers = decoded.Registers
-	p.Bytecode = decoded.Bytecode
-	p.Constants = constants
-	p.CatchTable = decoded.CatchTable
-	p.Params = decoded.Params
-	p.Metadata = decoded.Metadata
-
-	return nil
-}
 
 func encodeConstant(value runtime.Value) (constantJSON, error) {
 	if value == nil || value == runtime.None {
