+++ pkg/bytecode/program.go
@@ -1,6 +1,10 @@
 package bytecode
 
 import (
+	"fmt"
+
+	"github.com/goccy/go-json"
+
 	"github.com/MontFerret/ferret/v2/pkg/file"
 	"github.com/MontFerret/ferret/v2/pkg/runtime"
 )
@@ -25,3 +29,66 @@ type (
 		Metadata   Metadata
 	}
 )
+
+func (p *Program) MarshalJSON() ([]byte, error) {
+	if p == nil {
+		return []byte("null"), nil
+	}
+
+	constants := make([]constantJSON, len(p.Constants))
+
+	for i, value := range p.Constants {
+		encoded, err := encodeConstant(value)
+
+		if err != nil {
+			return nil, fmt.Errorf("bytecode.Program: encode constant %d: %w", i, err)
+		}
+
+		constants[i] = encoded
+	}
+
+	payload := programJSON{
+		Source:     p.Source,
+		Registers:  p.Registers,
+		Bytecode:   p.Bytecode,
+		Constants:  constants,
+		CatchTable: p.CatchTable,
+		Params:     p.Params,
+		Metadata:   p.Metadata,
+	}
+
+	return json.Marshal(payload)
+}
+
+func (p *Program) UnmarshalJSON(data []byte) error {
+	if p == nil {
+		return fmt.Errorf("bytecode.Program: UnmarshalJSON on nil pointer")
+	}
+
+	var decoded programJSON
+	if err := json.Unmarshal(data, &decoded); err != nil {
+		return err
+	}
+
+	constants := make([]runtime.Value, len(decoded.Constants))
+
+	for i, value := range decoded.Constants {
+		decodedValue, err := decodeConstant(value)
+
+		if err != nil {
+			return fmt.Errorf("bytecode.Program: decode constant %d: %w", i, err)
+		}
+
+		constants[i] = decodedValue
+	}
+
+	p.Source = decoded.Source
+	p.Registers = decoded.Registers
+	p.Bytecode = decoded.Bytecode
+	p.Constants = constants
+	p.CatchTable = decoded.CatchTable
+	p.Params = decoded.Params
+	p.Metadata = decoded.Metadata
+
+	return nil
+}
